+++ packages/terra-core-docs/src/terra-dev-site/doc/form-select/example/search/Invalid.jsx
@@ -12,7 +12,7 @@ const InvalidExample = () => {
   };
   return (
     <>
-      <SearchSelect placeholder="Select a color" isInvalid={isInvalid} ariaLabel={(isInvalid) ? 'Please select a color' : undefined} className={cx('form-select')} onSearch={handleOnChange} onChange={handleOnChange}>
+      <SearchSelect placeholder="Select a color" required isInvalid={isInvalid} ariaLabel={(isInvalid) ? 'Please select a color' : undefined} className={cx('form-select')} onSearch={handleOnChange} onChange={handleOnChange}>
         <SearchSelect.Option value="blue" display="Blue" />
         <SearchSelect.Option value="green" display="Green" />
         <SearchSelect.Option value="purple" display="Purple" />
@@ -20,6 +20,7 @@ const InvalidExample = () => {
         <SearchSelect.Option value="violet" display="Violet" />
       </SearchSelect>
       {isInvalid && <span className={cx('error-text')}>Please select a color</span>}
+      <p>Required: Please select a color from above list</p>
     </>
   );
 };