+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added new test example for `terra-form-search-select`
+  * Added instruction note for invalid example in `terra-search-select` and `terra-combo-box`.
 
 ## 1.53.0 - (December 11, 2023)